cnf ran: reorganize the gitopsztp helpers
* Reorganizes the gitopsztp helpers to make gitdetails and nodedelete their own modules
* Moves some functions into the ran/internal/ranhelper module for better reusability. The ranhelper package itself will be updated in the next PR.
* Removes some code that has been shifted to eco-goinfra
* Removes the argocd-operator dependency and shifts to the eco-goinfra types. This also allows for dumping the application CRs.
