+++ .golangci.yml
@@ -41,7 +41,6 @@ linters-settings:
         - "github.com/grafana-operator/grafana-operator"
         - "github.com/NVIDIA/gpu-operator"
         - "github.com/operator-framework/api"
-        - "github.com/argoproj-labs/argocd-operator/api"
         - "github.com/vmware-tanzu/velero/pkg/apis/velero/v1"
         - "github.com/vmware-tanzu/velero/pkg/generated/clientset/versioned/scheme"
         - "github.com/golang/glog"
