+++ tests/cnf/ran/gitopsztp/internal/gitdetails/gitdetails.go
@@ -1,4 +1,4 @@
-package helper
+package gitdetails
 
 import (
 	"context"
