+++ tests/cnf/ran/gitopsztp/internal/helper/helper.go
@@ -7,11 +7,16 @@ import (
 	"time"
 
 	"github.com/golang/glog"
-	"github.com/openshift-kni/eco-goinfra/pkg/argocd"
 	"github.com/openshift-kni/eco-goinfra/pkg/clients"
+	"github.com/openshift-kni/eco-goinfra/pkg/imageregistry"
 	"github.com/openshift-kni/eco-goinfra/pkg/ocm"
+	"github.com/openshift-kni/eco-goinfra/pkg/serviceaccount"
+	"github.com/openshift-kni/eco-goinfra/pkg/storage"
 	"github.com/openshift-kni/eco-gotests/tests/cnf/ran/gitopsztp/internal/tsparams"
+	"github.com/openshift-kni/eco-gotests/tests/cnf/ran/internal/ranhelper"
+	operatorv1 "github.com/openshift/api/operator/v1"
 	"k8s.io/apimachinery/pkg/util/wait"
+	configurationPolicyV1 "open-cluster-management.io/config-policy-controller/api/v1"
 )
 
 // WaitForPolicyToExist waits for up to the specified timeout until the policy exists.
@@ -38,72 +43,117 @@ func WaitForPolicyToExist(
 	return policy, err
 }
 
-// WaitForPolicyMessageToContainSubstring waits for the policy status message to contain the specified substring.
-func WaitForPolicyMessageToContainSubstring(client *clients.Settings, name, namespace, subString string) error {
-	glog.V(tsparams.LogLevel).Infof("Checking policy '%s' in namespace '%s' for message substring", name, namespace)
+// WaitForServiceAccountToExist waits for up to the specified timeout until the service account exists.
+func WaitForServiceAccountToExist(
+	client *clients.Settings, name, namespace string, timeout time.Duration) (*serviceaccount.Builder, error) {
+	var builder *serviceaccount.Builder
 
-	return wait.PollUntilContextTimeout(context.TODO(),
-		tsparams.ArgoCdChangeInterval, tsparams.ArgoCdChangeTimeout, true, func(ctx context.Context) (bool, error) {
-			policy, err := ocm.PullPolicy(client, name, namespace)
-			if err != nil {
-				return false, err
+	err := wait.PollUntilContextTimeout(
+		context.TODO(), tsparams.ArgoCdChangeInterval, timeout, true, func(ctx context.Context) (bool, error) {
+			var err error
+			builder, err = serviceaccount.Pull(client, name, namespace)
+
+			if err == nil {
+				return true, nil
 			}
 
-			details := policy.Definition.Status.Details
-			if len(details) > 0 && len(details[0].History) > 0 {
-				message := details[0].History[0].Message
-
-				glog.V(tsparams.LogLevel).Infof("Checking if message '%s' contains substring '%s'", message, subString)
-
-				return strings.Contains(message, subString), nil
+			if strings.Contains(err.Error(), "does not exist") {
+				return false, nil
 			}
 
-			return false, nil
+			return false, err
 		})
+
+	return builder, err
 }
 
-// WaitUntilSearchCollectorEnabled waits up to timeout until the KAC has the search collector addon enabled.
-func WaitUntilSearchCollectorEnabled(kac *ocm.KACBuilder, timeout time.Duration) error {
+// GetPolicyEvaluationIntervals is used to get the configured evaluation intervals for the specified policy.
+func GetPolicyEvaluationIntervals(policy *ocm.PolicyBuilder) (string, string, error) {
 	glog.V(tsparams.LogLevel).Infof(
-		"Waiting until search collector is enabled for KAC %s in namespace %s", kac.Definition.Name, kac.Definition.Namespace)
+		"Checking policy '%s' in namespace '%s' to fetch evaluation intervals",
+		policy.Definition.Name, policy.Definition.Namespace)
 
-	return wait.PollUntilContextTimeout(
-		context.TODO(), tsparams.ArgoCdChangeInterval, timeout, true, func(ctx context.Context) (bool, error) {
-			if !kac.Exists() {
-				glog.V(tsparams.LogLevel).Infof(
-					"KAC %s in namespace %s does not exist", kac.Definition.Name, kac.Definition.Namespace)
+	policyTemplates := policy.Definition.Spec.PolicyTemplates
+	if len(policyTemplates) < 1 {
+		return "", "", fmt.Errorf(
+			"could not find policy template for policy %s/%s", policy.Definition.Namespace, policy.Definition.Name)
+	}
 
-				return false, fmt.Errorf("kac %s in namespace %s does not exist", kac.Definition.Name, kac.Definition.Namespace)
-			}
+	configPolicy, err := ranhelper.UnmarshalRaw[configurationPolicyV1.ConfigurationPolicy](
+		policyTemplates[0].ObjectDefinition.Raw)
+	if err != nil {
+		return "", "", err
+	}
 
-			return kac.Definition.Spec.SearchCollectorConfig.Enabled, nil
-		})
+	complianceInterval := configPolicy.Spec.EvaluationInterval.Compliant
+	nonComplianceInterval := configPolicy.Spec.EvaluationInterval.NonCompliant
+
+	return complianceInterval, nonComplianceInterval, nil
 }
 
-// WaitForConditionInArgoCdApp waits up to timeout until the specified Argo CD app has a condition containing the
-// expectedMessage.
-func WaitForConditionInArgoCdApp(
-	client *clients.Settings, appName, namespace, expectedMessage string, timeout time.Duration) error {
+// RestoreImageRegistry restores the image registry with the provided name back to imageRegistryConfig, copying over the
+// labels, annotations, and spec from imageRegistryConfig, then waiting until the image registry is available again.
+func RestoreImageRegistry(
+	client *clients.Settings, imageRegistryName string, imageRegistryConfig *imageregistry.Builder) error {
+	currentImageRegistry, err := imageregistry.Pull(client, imageRegistryName)
+	if err != nil {
+		return err
+	}
+
+	if imageRegistryConfig.Definition.GetAnnotations() != nil {
+		currentImageRegistry.Definition.SetAnnotations(imageRegistryConfig.Definition.GetAnnotations())
+	}
+
+	if imageRegistryConfig.Definition.GetLabels() != nil {
+		currentImageRegistry.Definition.SetLabels(imageRegistryConfig.Definition.GetLabels())
+	}
+
+	currentImageRegistry.Definition.Spec = imageRegistryConfig.Definition.Spec
+
+	currentImageRegistry, err = currentImageRegistry.Update()
+	if err != nil {
+		return err
+	}
+
+	_, err = currentImageRegistry.WaitForCondition(operatorv1.OperatorCondition{
+		Type:   "Available",
+		Reason: "Removed",
+		Status: operatorv1.ConditionTrue,
+	}, tsparams.ArgoCdChangeTimeout)
+
+	return err
+}
+
+// CleanupImageRegistryConfig deletes the specified resources in the necessary order.
+func CleanupImageRegistryConfig(client *clients.Settings) error {
 	glog.V(tsparams.LogLevel).Infof(
-		"Checking application '%s' in namespace '%s' for condition with message '%s'", appName, namespace, expectedMessage)
+		"Cleaning up image registry resources with sc=%s, pv=%s, pvc=%s",
+		tsparams.ImageRegistrySC, tsparams.ImageRegistryPV, tsparams.ImageRegistryPVC)
 
-	return wait.PollUntilContextTimeout(
-		context.TODO(), tsparams.ArgoCdChangeInterval, timeout, true, func(ctx context.Context) (bool, error) {
-			app, err := argocd.PullApplication(client, appName, namespace)
-			if err != nil {
-				return false, err
-			}
+	// The resources must be deleted in the order of PVC, PV, then SC to avoid errors.
+	pvc, err := storage.PullPersistentVolumeClaim(client, tsparams.ImageRegistryPVC, tsparams.ImageRegistryNamespace)
+	if err == nil {
+		err = pvc.DeleteAndWait(tsparams.ArgoCdChangeTimeout)
+		if err != nil {
+			return err
+		}
+	}
 
-			for _, condition := range app.Definition.Status.Conditions {
-				if strings.Contains(condition.Message, expectedMessage) {
-					glog.V(tsparams.LogLevel).Info("Found matching condition")
+	pv, err := storage.PullPersistentVolume(client, tsparams.ImageRegistryPV)
+	if err == nil {
+		err = pv.DeleteAndWait(tsparams.ArgoCdChangeTimeout)
+		if err != nil {
+			return err
+		}
+	}
 
-					return true, nil
-				}
+	sc, err := storage.PullClass(client, tsparams.ImageRegistrySC)
+	if err == nil {
+		err = sc.DeleteAndWait(tsparams.ArgoCdChangeTimeout)
+		if err != nil {
+			return err
+		}
+	}
 
-				glog.V(tsparams.LogLevel).Infof("Condition message '%s' did not match", condition.Message)
-			}
-
-			return false, nil
-		})
+	return nil
 }
