+++ tests/cnf/ran/gitopsztp/internal/nodedelete/nodedelete.go
@@ -1,4 +1,4 @@
-package helper
+package nodedelete
 
 import (
 	"context"
@@ -11,15 +11,14 @@ import (
 	"github.com/openshift-kni/eco-goinfra/pkg/clients"
 	"github.com/openshift-kni/eco-goinfra/pkg/nodes"
 	"github.com/openshift-kni/eco-gotests/tests/cnf/ran/gitopsztp/internal/tsparams"
+	"github.com/openshift-kni/eco-gotests/tests/cnf/ran/internal/ranhelper"
 	. "github.com/openshift-kni/eco-gotests/tests/cnf/ran/internal/raninittools"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/util/wait"
 )
 
 // IsSnoPlusOne checks if the specified cluster has one control plane and one worker node.
 func IsSnoPlusOne(client *clients.Settings) (bool, error) {
-	controlPlanes, err := listNodesByLabel(client, RANConfig.ControlPlaneLabelMap)
+	controlPlanes, err := ranhelper.ListNodesByLabel(client, RANConfig.ControlPlaneLabelMap)
 	if err != nil {
 		return false, err
 	}
@@ -30,7 +29,7 @@ func IsSnoPlusOne(client *clients.Settings) (bool, error) {
 
 	glog.V(tsparams.LogLevel).Info("Exactly one control plane node found")
 
-	workers, err := listNodesByLabel(client, RANConfig.WorkerLabelMap)
+	workers, err := ranhelper.ListNodesByLabel(client, RANConfig.WorkerLabelMap)
 	if err != nil {
 		return false, err
 	}
@@ -54,7 +53,7 @@ func IsSnoPlusOne(client *clients.Settings) (bool, error) {
 
 // GetPlusOneWorkerName gets the name of the one worker in a SNO+1 cluster.
 func GetPlusOneWorkerName(client *clients.Settings) (string, error) {
-	workers, err := listNodesByLabel(client, RANConfig.WorkerLabelMap)
+	workers, err := ranhelper.ListNodesByLabel(client, RANConfig.WorkerLabelMap)
 	if err != nil {
 		return "", err
 	}
@@ -66,41 +65,6 @@ func GetPlusOneWorkerName(client *clients.Settings) (string, error) {
 	}
 
 	return "", fmt.Errorf("could not find a worker node for cluster")
-}
-
-// IsClusterStable checks if the provided cluster does not have any unschedulable nodes.
-func IsClusterStable(client *clients.Settings) (bool, error) {
-	nodeList, err := nodes.List(client)
-	if err != nil {
-		return false, err
-	}
-
-	for _, node := range nodeList {
-		if node.Definition.Spec.Unschedulable {
-			return false, nil
-		}
-	}
-
-	return true, nil
-}
-
-// WaitForNumberOfNodes waits up to timeout until the number of nodes on the cluster matches the expected.
-func WaitForNumberOfNodes(client *clients.Settings, expected int, timeout time.Duration) error {
-	return wait.PollUntilContextTimeout(
-		context.TODO(), 5*time.Second, timeout, true, func(ctx context.Context) (bool, error) {
-			nodeList, err := nodes.List(client)
-			if err != nil {
-				return false, err
-			}
-
-			if len(nodeList) == expected {
-				return true, nil
-			}
-
-			glog.V(tsparams.LogLevel).Infof("Expected %d nodes but found %d nodes", expected, len(nodeList))
-
-			return false, nil
-		})
 }
 
 // GetBmhNamespace returns the namespace for the specified BareMetalHost, if it exists.
@@ -138,35 +102,6 @@ func WaitForBMHDeprovisioning(client *clients.Settings, name, namespace string, 
 
 			return true, nil
 		})
-}
-
-// WaitForBMHAnnotation waits up to timeout until the specified annotation is present on the BMH.
-func WaitForBMHAnnotation(bareMetalHost *bmh.BmhBuilder, annotation string, timeout time.Duration) error {
-	return wait.PollUntilContextTimeout(
-		context.TODO(), tsparams.ArgoCdChangeInterval, timeout, true, func(ctx context.Context) (bool, error) {
-			if !bareMetalHost.Exists() {
-				glog.V(tsparams.LogLevel).Infof("BMH %s in namespace %s does not exist")
-
-				return false, fmt.Errorf(
-					"bmh object %s does not exist in namespace %s", bareMetalHost.Definition.Name, bareMetalHost.Definition.Namespace)
-			}
-
-			if _, ok := bareMetalHost.Object.Annotations[annotation]; !ok {
-				return false, nil
-			}
-
-			glog.V(tsparams.LogLevel).Infof(
-				"Annotation %s found for BMH %s/%s", annotation, bareMetalHost.Definition.Name, bareMetalHost.Definition.Namespace)
-
-			return true, nil
-		})
-}
-
-// listNodesByLabel returns a list of nodes that have the specified label.
-func listNodesByLabel(client *clients.Settings, labelMap map[string]string) ([]*nodes.Builder, error) {
-	return nodes.List(client, metav1.ListOptions{
-		LabelSelector: labels.Set(labelMap).String(),
-	})
 }
 
 // isNodeControlPlane checks whether the provided node is a control plane node.
